bugfix: fixup tsel case/lib

+++ test/tilelang_st/npu/a5/src/st/testcase/tsel/launch.cpp
@@ -13,71 +13,71 @@
 #endif
 
 // Case 0: f32 2x128
-extern "C" __global__ AICORE void TSEL_f32_2x128(__gm__ float *mask, __gm__ float *src0, __gm__ float *src1, __gm__ float *dst);
+extern "C" __global__ AICORE void TSEL_f32_2x128(__gm__ uint8_t *mask, __gm__ float *src0, __gm__ float *src1, __gm__ float *dst);
 
-void LaunchTSEL_f32_2x128(float *mask, float *src0, float *src1, float *dst, void *stream) {
-    TSEL_f32_2x128<<<1, nullptr, stream>>>((__gm__ float *)mask, (__gm__ float *)src0, (__gm__ float *)src1, (__gm__ float *)dst);
+void LaunchTSEL_f32_2x128(uint8_t *mask, float *src0, float *src1, float *dst, void *stream) {
+    TSEL_f32_2x128<<<1, nullptr, stream>>>((__gm__ uint8_t *)mask, (__gm__ float *)src0, (__gm__ float *)src1, (__gm__ float *)dst);
 }
 
 // Case 1: f32 2x32
-extern "C" __global__ AICORE void TSEL_f32_2x32(__gm__ float *mask, __gm__ float *src0, __gm__ float *src1, __gm__ float *dst);
+extern "C" __global__ AICORE void TSEL_f32_2x32(__gm__ uint8_t *mask, __gm__ float *src0, __gm__ float *src1, __gm__ float *dst);
 
-void LaunchTSEL_f32_2x32(float *mask, float *src0, float *src1, float *dst, void *stream) {
-    TSEL_f32_2x32<<<1, nullptr, stream>>>((__gm__ float *)mask, (__gm__ float *)src0, (__gm__ float *)src1, (__gm__ float *)dst);
+void LaunchTSEL_f32_2x32(uint8_t *mask, float *src0, float *src1, float *dst, void *stream) {
+    TSEL_f32_2x32<<<1, nullptr, stream>>>((__gm__ uint8_t *)mask, (__gm__ float *)src0, (__gm__ float *)src1, (__gm__ float *)dst);
 }
 
 // Case 2: f32 2x160
-extern "C" __global__ AICORE void TSEL_f32_2x160(__gm__ float *mask, __gm__ float *src0, __gm__ float *src1, __gm__ float *dst);
+extern "C" __global__ AICORE void TSEL_f32_2x160(__gm__ uint8_t *mask, __gm__ float *src0, __gm__ float *src1, __gm__ float *dst);
 
-void LaunchTSEL_f32_2x160(float *mask, float *src0, float *src1, float *dst, void *stream) {
-    TSEL_f32_2x160<<<1, nullptr, stream>>>((__gm__ float *)mask, (__gm__ float *)src0, (__gm__ float *)src1, (__gm__ float *)dst);
+void LaunchTSEL_f32_2x160(uint8_t *mask, float *src0, float *src1, float *dst, void *stream) {
+    TSEL_f32_2x160<<<1, nullptr, stream>>>((__gm__ uint8_t *)mask, (__gm__ float *)src0, (__gm__ float *)src1, (__gm__ float *)dst);
 }
 
 // Case 3: f32 2x512
-extern "C" __global__ AICORE void TSEL_f32_2x512(__gm__ float *mask, __gm__ float *src0, __gm__ float *src1, __gm__ float *dst);
+extern "C" __global__ AICORE void TSEL_f32_2x512(__gm__ uint8_t *mask, __gm__ float *src0, __gm__ float *src1, __gm__ float *dst);
 
-void LaunchTSEL_f32_2x512(float *mask, float *src0, float *src1, float *dst, void *stream) {
-    TSEL_f32_2x512<<<1, nullptr, stream>>>((__gm__ float *)mask, (__gm__ float *)src0, (__gm__ float *)src1, (__gm__ float *)dst);
+void LaunchTSEL_f32_2x512(uint8_t *mask, float *src0, float *src1, float *dst, void *stream) {
+    TSEL_f32_2x512<<<1, nullptr, stream>>>((__gm__ uint8_t *)mask, (__gm__ float *)src0, (__gm__ float *)src1, (__gm__ float *)dst);
 }
 
 // Case 4: f16 2x128
-extern "C" __global__ AICORE void TSEL_f16_2x128(__gm__ uint16_t *mask, __gm__ uint16_t *src0, __gm__ uint16_t *src1, __gm__ uint16_t *dst);
+extern "C" __global__ AICORE void TSEL_f16_2x128(__gm__ uint8_t *mask, __gm__ uint16_t *src0, __gm__ uint16_t *src1, __gm__ uint16_t *dst);
 
-void LaunchTSEL_f16_2x128(uint16_t *mask, uint16_t *src0, uint16_t *src1, uint16_t *dst, void *stream) {
-    TSEL_f16_2x128<<<1, nullptr, stream>>>((__gm__ uint16_t *)mask, (__gm__ uint16_t *)src0, (__gm__ uint16_t *)src1, (__gm__ uint16_t *)dst);
+void LaunchTSEL_f16_2x128(uint8_t *mask, uint16_t *src0, uint16_t *src1, uint16_t *dst, void *stream) {
+    TSEL_f16_2x128<<<1, nullptr, stream>>>((__gm__ uint8_t *)mask, (__gm__ uint16_t *)src0, (__gm__ uint16_t *)src1, (__gm__ uint16_t *)dst);
 }
 
 // Case 5: f16 2x32
-extern "C" __global__ AICORE void TSEL_f16_2x32(__gm__ uint16_t *mask, __gm__ uint16_t *src0, __gm__ uint16_t *src1, __gm__ uint16_t *dst);
+extern "C" __global__ AICORE void TSEL_f16_2x32(__gm__ uint8_t *mask, __gm__ uint16_t *src0, __gm__ uint16_t *src1, __gm__ uint16_t *dst);
 
-void LaunchTSEL_f16_2x32(uint16_t *mask, uint16_t *src0, uint16_t *src1, uint16_t *dst, void *stream) {
-    TSEL_f16_2x32<<<1, nullptr, stream>>>((__gm__ uint16_t *)mask, (__gm__ uint16_t *)src0, (__gm__ uint16_t *)src1, (__gm__ uint16_t *)dst);
+void LaunchTSEL_f16_2x32(uint8_t *mask, uint16_t *src0, uint16_t *src1, uint16_t *dst, void *stream) {
+    TSEL_f16_2x32<<<1, nullptr, stream>>>((__gm__ uint8_t *)mask, (__gm__ uint16_t *)src0, (__gm__ uint16_t *)src1, (__gm__ uint16_t *)dst);
 }
 
 // Case 6: f16 2x160
-extern "C" __global__ AICORE void TSEL_f16_2x160(__gm__ uint16_t *mask, __gm__ uint16_t *src0, __gm__ uint16_t *src1, __gm__ uint16_t *dst);
+extern "C" __global__ AICORE void TSEL_f16_2x160(__gm__ uint8_t *mask, __gm__ uint16_t *src0, __gm__ uint16_t *src1, __gm__ uint16_t *dst);
 
-void LaunchTSEL_f16_2x160(uint16_t *mask, uint16_t *src0, uint16_t *src1, uint16_t *dst, void *stream) {
-    TSEL_f16_2x160<<<1, nullptr, stream>>>((__gm__ uint16_t *)mask, (__gm__ uint16_t *)src0, (__gm__ uint16_t *)src1, (__gm__ uint16_t *)dst);
+void LaunchTSEL_f16_2x160(uint8_t *mask, uint16_t *src0, uint16_t *src1, uint16_t *dst, void *stream) {
+    TSEL_f16_2x160<<<1, nullptr, stream>>>((__gm__ uint8_t *)mask, (__gm__ uint16_t *)src0, (__gm__ uint16_t *)src1, (__gm__ uint16_t *)dst);
 }
 
 // Case 7: i8 2x128
-extern "C" __global__ AICORE void TSEL_i8_2x128(__gm__ int8_t *mask, __gm__ int8_t *src0, __gm__ int8_t *src1, __gm__ int8_t *dst);
+extern "C" __global__ AICORE void TSEL_i8_2x128(__gm__ uint8_t *mask, __gm__ int8_t *src0, __gm__ int8_t *src1, __gm__ int8_t *dst);
 
-void LaunchTSEL_i8_2x128(int8_t *mask, int8_t *src0, int8_t *src1, int8_t *dst, void *stream) {
-    TSEL_i8_2x128<<<1, nullptr, stream>>>((__gm__ int8_t *)mask, (__gm__ int8_t *)src0, (__gm__ int8_t *)src1, (__gm__ int8_t *)dst);
+void LaunchTSEL_i8_2x128(uint8_t *mask, int8_t *src0, int8_t *src1, int8_t *dst, void *stream) {
+    TSEL_i8_2x128<<<1, nullptr, stream>>>((__gm__ uint8_t *)mask, (__gm__ int8_t *)src0, (__gm__ int8_t *)src1, (__gm__ int8_t *)dst);
 }
 
 // Case 8: i8 2x32
-extern "C" __global__ AICORE void TSEL_i8_2x32(__gm__ int8_t *mask, __gm__ int8_t *src0, __gm__ int8_t *src1, __gm__ int8_t *dst);
+extern "C" __global__ AICORE void TSEL_i8_2x32(__gm__ uint8_t *mask, __gm__ int8_t *src0, __gm__ int8_t *src1, __gm__ int8_t *dst);
 
-void LaunchTSEL_i8_2x32(int8_t *mask, int8_t *src0, int8_t *src1, int8_t *dst, void *stream) {
-    TSEL_i8_2x32<<<1, nullptr, stream>>>((__gm__ int8_t *)mask, (__gm__ int8_t *)src0, (__gm__ int8_t *)src1, (__gm__ int8_t *)dst);
+void LaunchTSEL_i8_2x32(uint8_t *mask, int8_t *src0, int8_t *src1, int8_t *dst, void *stream) {
+    TSEL_i8_2x32<<<1, nullptr, stream>>>((__gm__ uint8_t *)mask, (__gm__ int8_t *)src0, (__gm__ int8_t *)src1, (__gm__ int8_t *)dst);
 }
 
 // Case 9: i8 2x160
-extern "C" __global__ AICORE void TSEL_i8_2x160(__gm__ int8_t *mask, __gm__ int8_t *src0, __gm__ int8_t *src1, __gm__ int8_t *dst);
+extern "C" __global__ AICORE void TSEL_i8_2x160(__gm__ uint8_t *mask, __gm__ int8_t *src0, __gm__ int8_t *src1, __gm__ int8_t *dst);
 
-void LaunchTSEL_i8_2x160(int8_t *mask, int8_t *src0, int8_t *src1, int8_t *dst, void *stream) {
-    TSEL_i8_2x160<<<1, nullptr, stream>>>((__gm__ int8_t *)mask, (__gm__ int8_t *)src0, (__gm__ int8_t *)src1, (__gm__ int8_t *)dst);
+void LaunchTSEL_i8_2x160(uint8_t *mask, int8_t *src0, int8_t *src1, int8_t *dst, void *stream) {
+    TSEL_i8_2x160<<<1, nullptr, stream>>>((__gm__ uint8_t *)mask, (__gm__ int8_t *)src0, (__gm__ int8_t *)src1, (__gm__ int8_t *)dst);
 }
